Rename MyMoney concent atoms
