+++ apps/store/src/features/carDealership/PayForTrial.tsx
@@ -14,8 +14,8 @@ import { useFormatter } from '@/utils/useFormatter'
 import { type TrialContract } from './carDealership.types'
 import { ConfirmPayWithoutExtensionButton } from './ConfirmPayWithoutExtensionButton'
 import { ExtensionOfferToggle } from './ExtensionOfferToggle'
+import { concentAcceptedAtom } from './MyMoneyConsent/myMoneyConcentAtom'
 import { MyMoneyConsent } from './MyMoneyConsent/MyMoneyConsent'
-import { concentAcceptedAtom } from './MyMoneyConsent/MyMoneyConsentAtom'
 import { PriceBreakdown } from './PriceBreakdown'
 import { ProductItemContractContainerCar } from './ProductItemContractContainer'
 import { type MyMoneyConsentProps } from './TrialExtensionForm/TrialExtensionForm'
