+++ apps/store/src/features/carDealership/TrialExtensionForm/TrialExtensionForm.tsx
@@ -20,10 +20,10 @@ import { PageLink } from '@/utils/PageLink'
 import { useFormatter } from '@/utils/useFormatter'
 import { type TrialContract } from '../carDealership.types'
 import { ExtensionOfferToggle } from '../ExtensionOfferToggle'
+import { concentAcceptedAtom } from '../MyMoneyConsent/myMoneyConcentAtom'
 import { MyMoneyConsent } from '../MyMoneyConsent/MyMoneyConsent'
-import { concentAcceptedAtom } from '../MyMoneyConsent/MyMoneyConsentAtom'
+import { myMoneyConsentConfirmationFlagAtom } from '../MyMoneyConsent/MyMoneyConsentConfirmation/myMoneyConcentConfirmationAtom'
 import { MyMoneyConsentConfirmation } from '../MyMoneyConsent/MyMoneyConsentConfirmation/MyMoneyConsentConfirmation'
-import { myMoneyConsentConfirmationFlagAtom } from '../MyMoneyConsent/MyMoneyConsentConfirmation/MyMoneyConsentConfirmationAtom'
 import { PriceBreakdown } from '../PriceBreakdown'
 import { ProductItemContractContainerCar } from '../ProductItemContractContainer'
 import { EditActionButton } from './EditActionButton'
